Rename contentstore/views/requests to helpers
This has caused confusion, since this file gets imported when trying to
import the requests library.  And besides, how was this file about
requests in the first place?  Helpers is a much better name.
