+++ cms/djangoapps/contentstore/views/__init__.py
@@ -9,13 +9,13 @@ from .checklist import *
 from .component import *
 from .course import *
 from .error import *
+from .helpers import *
 from .item import *
 from .import_export import *
 from .preview import *
 from .public import *
 from .user import *
 from .tabs import *
-from .requests import *
 try:
     from .dev import *
 except ImportError:
